Разбивает код генерации фото по модулям
