+++ js/main.js
@@ -1,65 +1,3 @@
-const PICTURE_COUNT = 25;
-const LIKE_MIN_COUNT = 15;
-const LIKE_MAX_COUNT = 200;
-const AVATAR_COUNT = 6;
-const COMMENT_COUNT = 6;
-const COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const DESCRIPTIONS = [
-  'Фотография не является отражением реальности. Она есть реальность этого отражения.',
-  'Реальность становится всё больше похожа на фотографии.',
-  'В любой ситуации всегда улыбайтесь.',
-  'В простоте есть удивительная красота.',
-  'Думать следует до и после съёмки, никогда во время её.',
-  'На всякий случай, а то вдруг вы забыли, как я выгляжу…'
-];
-const NAMES = ['Иван', 'Мария', 'Кристоф', 'Виктор', 'Юлия', 'Люпита'];
+import {createPhoto} from './data.js';
 
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-function createIdGenerator () {
-  let lastGeneratedId = 0;
-
-  return function () {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-}
-
-const generateCommentId = createIdGenerator();
-const generatePhotoId = createIdGenerator();
-const generateUrlId = createIdGenerator();
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
-  message: getRandomArrayElement(COMMENTS),
-  name: getRandomArrayElement(NAMES)
-});
-
-console.log(createComment());
-
-const getPhoto = () => ({
-  id: generatePhotoId(),
-  url: `photos/${generateUrlId()}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
-  comments: Array.from({length: getRandomInteger(1, COMMENT_COUNT)}, createComment)
-});
-
-const similarPhoto = Array.from({length: PICTURE_COUNT}, getPhoto);
-
-console.log(similarPhoto);
+createPhoto();
